Remove stale extension files
